fix(seq_orders): bug found in seq_orders function
If function argument for model did not have a matching model in the
input file, then the last line read would be returned as the results of
the file processing. Now, the function will return empty lists or
strings depending on the output type. Also, only processs full line if
the model name and argument match to save on extra processing.